Improve overview performance by calculating timestamp outside iterator
What?
=====

For repositories with a large number of commits, filtering to the most
recent six months when that calculation occurs within the iterator can
be time-consuming.

This shifts the calculation of the six-month window outside of the
iterator to reduce the number of times we calculate this value to one.

This results in a ~30% speed-up.