Introduce `churn` subcommand
What?
=====

This introduces a rudimentary file churn calculation by traversing diffs
and, based on the diff status, tracks additions, removals, and
modifications of each path.

This currently only accounts for a subset of diff delta statuses,
notably added, removed, and modified statuses. As far as I can tell,
libgit2 does not correctly report the renamed status correctly even in
the simplest of cases (e.g. using `git mv current_path new_path`).
Currently, this is reported as two separate deltas, one with "added"
status and one with "deleted" status.

In the future, it would be beneficial to support this case by
effectively replacing the old `PathBuf` within the `BTreeMap` with the
new `PathBuf`, maintaining changes count.